Fixups for sparc64
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@26558 72102866-910b-0410-8b05-ffd578937521

+++ Tools/gdnc.m
@@ -242,7 +242,7 @@ ihandler(int sig)
 @interface	GDNCObserver : NSObject
 {
 @public
-  unsigned		observer;
+  unsigned long		observer;
   NSString		*notificationName;
   NSString		*notificationObject;
   NSString		*selector;
